+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Enhancements
 
 * Added `to_int` method for characters.
 * `Token::FloatConstant` and `Token::DecimalConstant` now carry the original text representation for use in, say, a _token mapper_.
+* `Dynamic::is_fnptr` is made a public API.
 
 
 Version 1.16.2